3.6.1.0.06

+++ phpshop/admpanel/modules/action.php
@@ -31,41 +31,54 @@ function GetModuleInfo($name) {
     }
 }
 
-switch($_POST['moduleAction']) {
-    case("off"):
-    
-        PHPShopObj::loadClass("modules");
-        $PHPShopModules = new PHPShopModules($_classPath."modules/");
-        $ModValue=$PHPShopModules->getModValue();
-        
-        // Удаление базы
-        if(is_array($ModValue['base'][$_POST['moduleId']]))
-            foreach($ModValue['base'][$_POST['moduleId']] as $val) mysql_query("DROP TABLE ".$val);
-        
-        // Удаление полей
-        if(is_array($ModValue['field'][$_POST['moduleId']]))
-            foreach($ModValue['field'][$_POST['moduleId']] as $key=>$val) mysql_query("ALTER TABLE `".$val."` DROP `".$key."` ");
-        
-        $sql="delete from ".$GLOBALS['SysValue']['base']['modules']." where path='".$_POST['moduleId']."'";
-        $result=mysql_query($sql);
-        break;
-    
-    case("on"):
-    // Информация по модулю
-        $Info = GetModuleInfo($_POST['moduleId']);
-        $sql="INSERT INTO ".$GLOBALS['SysValue']['base']['modules']."  VALUES ('".$_POST['moduleId']."','".$Info ['name']."','".date("U")."')";
-        $result=mysql_query($sql);
-        
-        // Устанавливаем БД модуля
-        $modulePath="../../modules/".$_POST['moduleId']."/install/module.sql";
-        if(file_exists($modulePath)) {
-            $moduleSQLFile=file_get_contents($modulePath);
-            $SQLArray=explode(";",$moduleSQLFile);
-            foreach($SQLArray as $val) $result=mysql_query($val);
-        }
-        break;
-    
-}
+if(!CheckedRules($UserStatus["module"],1))
+    switch($_POST['moduleAction']) {
+        case("off"):
+
+            PHPShopObj::loadClass("modules");
+            $PHPShopModules = new PHPShopModules($_classPath."modules/");
+            $ModValue=$PHPShopModules->getModValue();
+
+            // Удаление базы
+            if(is_array($ModValue['base'][$_POST['moduleId']]))
+                foreach($ModValue['base'][$_POST['moduleId']] as $val) mysql_query("DROP TABLE ".$val);
+
+            // Удаление полей
+            if(is_array($ModValue['field'][$_POST['moduleId']]))
+                foreach($ModValue['field'][$_POST['moduleId']] as $key=>$val) mysql_query("ALTER TABLE `".$val."` DROP `".$key."` ");
+
+            $sql="delete from ".$GLOBALS['SysValue']['base']['modules']." where path='".$_POST['moduleId']."'";
+            $result=mysql_query($sql);
+            break;
+
+        case("on"):
+        // Информация по модулю
+            $Info = GetModuleInfo($_POST['moduleId']);
+
+            if(empty($Info['trial'])) $date_end=0;
+            else  $date_end=time()+2592000;
+
+            if(empty($Info['key'])) $key=null;
+            else $key=$Info['key'];
+
+            $sql="INSERT INTO ".$GLOBALS['SysValue']['base']['modules']."  VALUES ('".$_POST['moduleId']."','".$Info ['name']."',".time().")";
+            $result=mysql_query($sql);
+
+            if(!empty($date_end))
+            $sql="INSERT INTO ".$GLOBALS['SysValue']['base']['modules_key']."  VALUES ('".$_POST['moduleId']."',".$date_end.",'".$key."','".md5($_POST['moduleId'].$date_end.$_SERVER['SERVER_NAME'].$key)."')";
+            $result=mysql_query($sql);
+
+            // Устанавливаем БД модуля
+            $modulePath="../../modules/".$_POST['moduleId']."/install/module.sql";
+            if(file_exists($modulePath)) {
+                $moduleSQLFile=file_get_contents($modulePath);
+                $SQLArray=explode(";",$moduleSQLFile);
+                foreach($SQLArray as $val) $result=mysql_query($val);
+            }
+
+            break;
+
+    }
 
 
 $_RESULT = array(
